fixes #4
Fixes #4 the home widget had to be put into a seperate class otherwise exceptions were raised.
